Occupied component styling broken up and extracted to components
Repeated styles inside of the occupied component were removed
Styles unique to other components than occupied were extracted
into files associated with folders of those componenets
