+++ app/components/ApplicationsCard/styles.scss
@@ -204,3 +204,45 @@
     height: 55px;
   }
 }
+
+.cardLine {
+  background-color: $grey;
+  width: 85%;
+  border: none;
+  height: 1px;
+  margin-top: 20px;
+  margin-bottom: 15px;
+}
+
+.cardFooter {
+  width: 90%;
+  display: flex;
+  height: 20px;
+  align-items: center;
+  margin-left: 18px;
+}
+
+#cardFooterText {
+  color: $icon;
+  font-size: 11px;
+  margin: 0;
+  margin-left: 10px;
+}
+
+.cardFooterIcon {
+  color: $icon;
+  font-size: 14px;
+  margin: 0;
+}
+
+#databaseName {
+  margin-top: 14px;
+  margin-bottom: 0;
+  font-size: 40px;
+  width: 280px;
+}
+
+#serviceName {
+  font-size: 11px;
+  margin-top: 6px;
+}
