+++ app/components/FirstLaunch/FirstLaunch.scss
