+++ app/components/Occupied/Occupied.scss
@@ -51,186 +51,12 @@
   }
 }
 
-.navIcon {
-  font-size: 34px;
-  vertical-align: middle;
-  text-align: center;
-  color: $icon;
-  background-color: transparent;
-  box-sizing: content-box;
-  border-radius: 50%;
-  line-height: 1.5;
-  white-space: nowrap;
-  outline: 0;
-  border: none;
-  cursor: pointer;
-  position: relative;
-  min-height: auto;
-  white-space: nowrap;
-  margin: 5px;
-  margin-top: 15px;
-  padding: 5px;
-}
-
-#searchIcon {
-  font-size: 26px;
-  font-weight: 400px;
-  padding: 8px;
-  margin: 0px;
-  margin-left: 15px;
-  margin-right: 15px;
-
-  &:hover {
-    cursor: pointer;
-    box-shadow: $boxshadow3;
-    color: $gblue;
-  }
-}
-
-// Dashboard Icon
-#dashboardIcon {
-  margin-top: 4px;
-  margin-left: -35px;
-  font-size: 26px;
-  padding: 8px;
-
-  &:hover {
-    cursor: pointer;
-    box-shadow: $boxshadow3;
-    // background-color: rgba(255, 255, 255, 0.3);
-    background-color: $background;
-    color: $ggreen;
-  }
-}
-
-.dashboardIconArea .dashboardTooltip::after {
-  content: '';
-  position: absolute;
-  bottom: 50px;
-  left: 35px;
-  border-width: 5px;
-  border-style: solid;
-  border-color: transparent transparent $background transparent;
-}
-
-// Notification Icon
-#notificationsIcon {
-  margin-top: 4px;
-  margin-left: 2px;
-  font-size: 26px;
-
-  // padding: 8px;
-  &:hover {
-    cursor: pointer;
-    box-shadow: $boxshadow3;
-    // background-color: rgba(255, 255, 255, 0.3);
-    background-color: $background;
-    color: $gred;
-  }
-}
-
-.notificationsIconArea {
-  position: relative;
-  display: inline-block;
-}
-
-.notificationsIconArea:hover .notificationsTooltip {
-  visibility: visible;
-}
-
-.notificationsIconArea .notificationsTooltip {
-  visibility: hidden;
-  width: 80px;
-  background-color: $background;
-  color: $gred;
-  text-align: center;
-  border-radius: 6px;
-  padding: 5px 0;
-  font-size: 12px;
-  /* Position the tooltip */
-  position: absolute;
-  z-index: 1;
-  top: 50px;
-  right: -14px;
-}
-
-.notificationsIconArea .notificationsTooltip::after {
-  content: '';
-  position: absolute;
-  bottom: 37px;
-  left: 35px;
-  border-width: 5px;
-  border-style: solid;
-  border-color: transparent transparent $background transparent;
-}
-
-// Person Icon
-#personIcon {
-  margin-top: 4px;
-  margin-left: 2px;
-  font-size: 26px;
-
-  // padding: 8px;
-  &:hover {
-    cursor: pointer;
-    box-shadow: $boxshadow3;
-    // background-color: rgba(255, 255, 255, 0.3);
-    background-color: $background;
-    color: $gyellow;
-  }
-}
-
-.personIconArea {
-  position: relative;
-  display: inline-block;
-}
-
-.personIconArea:hover .personTooltip {
-  visibility: visible;
-}
-
-.personIconArea .personTooltip {
-  // visibility: hidden;
-  // width: 80px;
-  // background-color: $background;
-  color: $gyellow;
-  text-align: center;
-  // border-radius: 6px;
-  // padding: 5px 0 !important;
-  // font-size: 12px;
-  // /* Position the tooltip */
-  // position: absolute;
-  // z-index: 1;
-  // top: 63px;
-  // right: 10px;
-}
-
-// .personIconArea .personTooltip::after {
-//   content: '';
-//   position: absolute;
-//   bottom: 37px;
-//   left: 35px;
-//   border-width: 5px;
-//   border-style: solid;
-//   border-color: transparent transparent $background transparent;
-// }
-
-// .dashboardIconWrapper {
-//   margin-right: 37px;
-// }
-
-.btns {
-  display: flex;
-  justify-content: space-around;
-}
-
 button {
   background-color: transparent;
   border: 0;
   outline: none;
   white-space: nowrap;
   margin: 10px 0;
-  // padding: 0px;
   color: transparent;
 
   &:hover {
@@ -253,20 +79,6 @@ button {
     filter: none;
     cursor: pointer;
   }
-}
-
-#textInput {
-  background-color: transparent;
-  font-family: 'Roboto';
-  font-size: 16px;
-  outline: none;
-  border: 0;
-  padding: 2px;
-  // transform: translateY(-50%);
-  transition: all 0.25s ease;
-  pointer-events: none;
-  text-align: left;
-  margin-top: 5px;
 }
 
 input:-webkit-autofill,
@@ -292,46 +104,4 @@ select:-webkit-autofill:focus {
   align-self: center;
   width: 80%;
   margin-top: 50px;
-}
-
-.cardLine {
-  background-color: $grey;
-  width: 85%;
-  border: none;
-  height: 1px;
-  margin-top: 20px;
-  margin-bottom: 15px;
-}
-
-.cardFooter {
-  width: 90%;
-  display: flex;
-  height: 20px;
-  align-items: center;
-  margin-left: 18px;
-}
-
-#cardFooterText {
-  color: $icon;
-  font-size: 11px;
-  margin: 0;
-  margin-left: 10px;
-}
-
-.cardFooterIcon {
-  color: $icon;
-  font-size: 14px;
-  margin: 0;
-}
-
-#databaseName {
-  margin-top: 14px;
-  margin-bottom: 0;
-  font-size: 40px;
-  width: 280px;
-}
-
-#serviceName {
-  font-size: 11px;
-  margin-top: 6px;
 }
