+++ app/containers/LandingPageContainer.tsx
@@ -1,7 +1,7 @@
 import React, { useContext } from 'react';
 import { redirect } from 'react-router-dom';
 import CreateAdmin from '../components/CreateAdmin';
-import FirstLaunch from '../components/FirstLaunch';
+import FirstLaunch from '../components/FirstLaunch/FirstLaunch';
 import Login from '../components/Login';
 import SignUp from '../components/SignUp';
 import { DashboardContext } from '../context/DashboardContext';
